auto_trader/requirements.txt
python-telegram-bot==12.7
pandas==1.2.2
requests==2.26.0
matplotlib==3.4.2
mplfinance==0.12.7a17
ccxt==1.54.54
bybit==0.2.12
certifi==2021.5.30
click==7.1.2
Flask==1.1.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
Werkzeug==1.0.1
wincertstore==0.2
tensorflow==2.11.0
gym==0.21
